assignment/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
Django==3.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
idna==2.8
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.3
kiwisolver==1.1.0
lml==0.0.9
MarkupSafe==1.1.1
matplotlib==3.1.2
numpy==1.17.4
openpyxl==2.5.14
pandas==0.25.3
pyexcel==0.5.15
pyexcel-io==0.5.20
pyexcel-xlsx==0.5.7
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
texttable==1.6.2
uritemplate==3.0.0
urllib3==1.25.7
